Time - Greeting Widgets
Just pushing this to clear up my github a little. This will be the first of ~3 pushes for the basic version of the new home screen thats focused on creating standards. This one pushes the date and time widget and the greeting widget, and guts the css for the weather widget. I think once i get the basics over from figma/zeplin into the code I am going to start on the standardizing of formats so we can easily replicate basic widget shells moving forward (size, coloring, font, etc) for the section and frame classes. Hopefully will get the next two pushes out tomorrow and Sunday.