Greeting Block
Builds the basic ground work of a greeting component. Uses the time functionality from the timeBlock.js file to greet the user based on if 'PM' or 'AM'. Will be able to pull user name from the database as well when fixed. CSS uses roboto, which will be standard for text.